+++ ompi/mca/mpool/base/mpool_base_mem_cb.c
@@ -55,9 +55,9 @@ void mca_mpool_base_mem_cb(void* base, size_t size, void* cbdata,
                     base, size);
 
             if(rc != OMPI_SUCCESS && true == warn) {
-                 opal_output(0, "Memory %p:%llu cannot be freed from the "
+                 opal_output(0, "Memory %p:%lu cannot be freed from the "
                          "registration cache. Possible memory corruption.\n",
-                         base, size);
+                         base, (unsigned long)size);
                  warn = false;
             }
         }
